requirements.txt
attrs==17.4.0
cycler==0.10.0
matplotlib==2.1.2
numpy==1.14.1
pdfminer3k==1.3.1
pluggy==0.6.0
ply==3.11
py==1.5.2
pyparsing==2.2.0
pytest==3.4.1
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
